requirements.txt
ffmpeg-python
python-dotenv
pytubefix
openai-whisper